[api] Complete /api/containers/delete

+++ database/delete.go
@@ -4,15 +4,20 @@ import (
 	"log"
 )
 
-func DeleteContainer(containerName string) bool {
+func DeleteContainer(containerId string) bool {
+	if flag := CheckContainerExists(containerId); !flag {
+		log.Printf("Container %v not exist", containerId)
+		return false
+	}
+
 	stmt, err := GetDBInstance().Prepare("DELETE FROM container WHERE container_id = ?")
 	if err != nil {
 		log.Panicln(err)
 	}
 
-	_, err = stmt.Exec(containerName)
+	_, err = stmt.Exec(containerId)
 	if err != nil {
-		return false
+		log.Panicln(err)
 	}
 
 	return true
